:bug: Fix bugs in the app
Fix bug with the routes on the client in the production app. Fix bug
with categories in the navigation when there is no connection with
server. Fix bug with matching passwords in the Profil component. Fix
getBoundingClientRect bug in the admin views. Fix bug with skip and
limit in the Products component. Fix bug with products in the cart and
last viewed products in the local storage. Add button to fetch more
comments. Add more regexps. Add default parameters skip and limit in
the services. Add redirect to the 404 page when user go to the product
that not exist.